Added some documentation to page data and fixed re-mounting bug in narrative page.
